support reading from multiple iovec entries
Previously we were only ever reading from the first iovec in the msghdr,
which resulted in data sometimes being dropped. This changes the eBPF
code to emit a new entry for each iovec in the msghdr, up to a max of 10
entries per traced call. While it's still possible that we'll miss out
on data if >10 iovecs are passed to a single sendmsg call, this should
make it less likely to happen under normal circumstances.

Adds a test case that exercises this logic by writing several packets at
once to the same socket.