sort incoming events
To provide a better guarantee on the ordering of events, run all events
through a sorting algorithm that allows for a 100ms window in which
events may come in out of order. It ensures that as long as for any
event A, all events that preceded A in kernel time arrive in userspace
within 100ms, the order of events written to stdout should be correct.